Fix nonHtml issues with extensions containing #s
nonHtml was short-circuiting and returning false if the link extension contained a number, such as .mp3. 

The result of this behavior is that all .mp3 links would automatically be processed using xhr and in some instances present the user with a HTTP 406 depending on how the client handled the request.